olap/requirements.txt
psycopg2-binary==2.9.9
pytest==8.3.4
python-dotenv==1.0.1
